automation/requirements.txt
google-generativeai>=0.8.0
Pillow>=10.2.0
supabase>=2.3.4
feedparser>=6.0.10
python-dotenv>=1.0.1
requests>=2.31.0
google-api-python-client>=2.118.0
google-auth-oauthlib>=1.2.0
google-auth>=2.28.1
httpx>=0.27.0
tenacity>=8.2.3
